requirements.txt
openai
faiss-cpu
flask
twilio
python-dotenv
numpy
